alterando viw do editar postagem. a pagina mostra a postagem pelo id no template corretamente, mas ainda não consegue salvar e atualizar no formulario do django.

+++ post/views.py
@@ -1,10 +1,11 @@
 from django.http import Http404
-from django.shortcuts import redirect, render
+from django.shortcuts import get_object_or_404, redirect, render
 from rest_framework.viewsets import ModelViewSet
 from .serializers import PostSerializer  
 from .models import Post as criarPost
 from post.forms import PostForm, ComentarioForm
 from categorias.models import Categoria
+from inicio.views import inicio
 
 
 # Create your views here.
@@ -25,19 +26,37 @@ def gerenciar_postagem(request):
 
 
 def editar_postagem(request, pk):
+    template_name = 'galeria/editar_postagem.html'
+    categorias = Categoria.objects.all()
     postagem = criarPost.objects.get(pk=pk)
-    form = PostForm(request.POST or None, instance=postagem)
+    form = PostForm(instance=postagem)
     sucesso = False
-    if form.is_valid():
-        form.save()
-        sucesso = True
+    if request.method == 'POST':
+        form = PostForm(request.FILES or None, request.POST or None, instance=postagem)
+        print(postagem.pk)
+        
+        if form.is_valid():
+            postagem = form.save(False)
+            postagem.usuario = request.user
+            postagem.slug = form.cleaned_data['postagem.titulo'].replace(' ', '-').lower()
+            postagem = criarPost.objects.create(usuario=postagem.usuario, slug=postagem.slug, titulo=form.cleaned_data['titulo'], descricao=form.cleaned_data['descricao'], texto=form.cleaned_data['texto'], imagem=form.cleaned_data['imagem'], categoria=form.cleaned_data['categoria'])
+            postagem.save()
+            sucesso = True
+            return redirect('postagem_detalhada', pk=postagem.pk)
+        else:
+            print(form.errors)
+            raise Http404
+    else:
+            form = PostForm(instance=postagem)
     contexto = {
         'postagem': postagem,
         'form': form,
-        'sucesso': sucesso
+        'sucesso': sucesso,
+        'categorias': categorias
     }
-    return render(request, 'galeria/criar_postagem.html', contexto)
-
+    return render(request, template_name, contexto)
+ 
+ 
 def criar_postagem(request):
     template_name = 'galeria/criar_postagem.html'
     categorias = Categoria.objects.all()
@@ -52,7 +71,7 @@ def criar_postagem(request):
             post.slug = form.cleaned_data['titulo'].replace(' ', '-').lower()
             post.save()
             sucesso = True
-            return redirect('criar_postagem')
+            return redirect('postagem_detalhada', pk=post.pk)
         else:
             print(form.errors)
             raise Http404
@@ -70,8 +89,9 @@ def criar_postagem(request):
 
 def excluir_postagem(request, pk):
     postagem = criarPost.objects.get(pk=pk)
-    postagem.delete()
-    return render(request, 'galeria/postagem.html')
+    if request.method == 'POST':
+        postagem.delete()
+    return render(request, 'galeria/postagemdousuario.html')
 
 def addcomentario (request, pk):
     postagem = criarPost.objects.get(pk=pk)
